authentataion

+++ controller/allExpense.js
@@ -1,15 +1,16 @@
 const Expense=require('../models/expense')
 exports.addExpenseAmount=async(req,res)=>{
-// try{}catch(e){
-//     console.log('error while adding item')
-// }
+try{
     const{amount,comment,catagory}=req.body
-   const exp=await Expense.create({amount:amount,comment:comment,catagory:catagory})
+   const exp=await req.user.createExpense({amount:amount,comment:comment,catagory:catagory})
    res.status(200).json(exp)
+}catch(e){
+    console.log('error while adding item')
+}
 }
 exports.deleteAmount=async(req,res)=>{
     try{
-   await Expense.destroy({where:{id:req.params.id}})
+   await Expense.destroy({where:{id:req.params.id,UserId:req.user.id}})
    res.status(200).json({success:true})
 }catch(err){
     res.status(400).json({message:'something went wrong'})
@@ -17,16 +18,16 @@ exports.deleteAmount=async(req,res)=>{
 }
 exports.ediiAmount=async(req,res)=>{
     try{
-       const exp = await Expense.findOne({where:{id:req.params.id}})
-        res.status(200).json(exp)
+        const exp = await req.user.getExpenses({where:{id:req.params.id}})
+            res.status(200).json(exp)
      }catch(err){
          res.status(400).json({message:'something went wrong'})
      }
 }
 exports.showAllExpense=async(req,res)=>{
     try{
-   const exp= await Expense.findAll()
-   res.status(200).json(exp)
+        const exp= await req.user.getExpenses()
+        res.status(200).json(exp)
     }catch(err){
     res.status(400).json({message:'something went wrong'})
     }
